requirements.txt
cx_oracle
django-rest-framework